+++ test/test_mpi.py
@@ -83,9 +83,6 @@ def test_observation_time():
 def test_construction_from_detectors():
     comm_world = lbs.MPI_COMM_WORLD
 
-    if comm_world.rank == 0:
-        print(f"MPI configuration: {lbs.MPI_CONFIGURATION}")
-
     det1 = dict(
         name="pol01",
         wafer="mywafer",
@@ -649,6 +646,39 @@ def __run_test_in_same_folder(test_fn: Callable) -> None:
             print("Failure", file=stderr)
 
 
+def test_nullify_mpi(tmp_path):
+    start_time = 0
+    time_span_s = 2
+    sampling_hz = 12
+
+    sim = lbs.Simulation(
+        base_path=tmp_path,
+        start_time=start_time,
+        duration_s=time_span_s,
+        random_seed=12345,
+    )
+
+    det = lbs.DetectorInfo(
+        name="Dummy detector",
+        sampling_rate_hz=sampling_hz,
+        bandcenter_ghz=100.0,
+        quat=[0.0, 0.0, 0.0, 1.0],
+    )
+
+    num_of_obs = 12
+    sim.create_observations(detectors=[det], num_of_obs_per_detector=num_of_obs)
+
+    # Assert TOD is initially non-zero and specific to rank
+    for obs in sim.observations:
+        obs.tod[:, :] = lbs.MPI_COMM_WORLD.rank + 1
+
+    sim.nullify_tod()
+
+    # Assert TOD is now zero
+    for obs in sim.observations:
+        assert np.all(obs.tod == 0)
+
+
 if __name__ == "__main__":
     test_functions = [
         test_observation_time,
@@ -668,6 +698,7 @@ if __name__ == "__main__":
     same_folder_test_functions = [
         test_write_hdf5_mpi,
         test_issue314,
+        test_nullify_mpi,
     ]
 
     for cur_test_fn in same_folder_test_functions: